+++ src/Organisms/App/App.test.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from 'src/Organisms';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
+
   ReactDOM.render(<App />, div);
   ReactDOM.unmountComponentAtNode(div);
 });
